Move tests back to pps/cmds.
